pyproject.toml
[project]
name = "genai-llms"
version = "0.1.0"
description = "EN->JA translation with baseline, fine-tuning, RAG, and agentic RAG."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "modal>=1.1.0",
  "huggingface-hub>=0.32.0",
  "boto3>=1.37.0",
  "pyyaml>=6.0.2",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["modal_jobs*", "src*"]
